requirements.txt
accelerate==0.21.0
apex==0.9.10dev
decord==0.6.0
einops==0.7.0
fairscale==0.4.4
flash_attn==2.0.8
importlib_metadata==6.8.0
iopath==0.1.10
matplotlib==3.7.2
numpy==1.25.2
omegaconf==2.3.0
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
packaging==23.2
pandas==1.5.3
Pillow==9.5.0
PyYAML==6.0.1
Requests==2.31.0
salesforce_lavis==1.0.1
scikit_learn==1.2.1
scipy==1.11.3
sentencepiece==0.1.99
skimage==0.0
spacy==3.6.1
tensorflow==2.14.0
timm==0.4.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.2
transformers==4.31.0
wandb==0.15.8
webdataset==0.2.48
xformers==0.0.22.post4
